Eliminate v4 wire model layer, fix v5 test failures
Refactor all v4 packet models to use @ProtocolMessage annotations directly
on domain classes, removing the intermediate wire/ layer (7 files deleted).
ACK packets use generated codecs, Subscribe/Unsubscribe use @RemainingBytes,
Publish/Connect inline encode/decode (flags in byte1).

Fix 65 v5 test failures caused by buffer-codec removing sizeOf API:
refactor Auth, Disconnect, ConnAck to use @ProtocolMessage body classes
with @WhenRemaining for variable-length fields instead of WireEncoded.

Remove sizeOfFunction from codec-spi providers (removed from buffer API).

+++ benchmark/src/commonMain/kotlin/com/ditchoom/mqtt/benchmark/WritePathBenchmark.kt
@@ -63,7 +63,7 @@ class WritePathBenchmark {
 
     @Benchmark
     fun serializePuback(bh: Blackhole) {
-        val puback = PublishAcknowledgment(1)
+        val puback = PublishAcknowledgment(1u.toUShort())
         val buf = puback.toBuffer()
         bh.consume(buf)
     }
